analysis/requirements.txt
pytesseract==0.3.7
pyspellchecker==0.6.2
flair==0.8.0.post1
pandas==1.2.4
scipy==1.6.2
word2number==1.1
numpy==1.19.5
opencv_python==4.5.2.54
pdf2image==1.15.1
Pillow==8.3.2
spellchecker==0.4
